k8s: fix issue with SSL certs not renewing.
Fix an issue with SSL certificates not renewing because of a configuration
error. In cluster-issuer.yaml, we refer to a secret called
`cloudflare-api-token`. This secret is generated by a secretGenerator in
kustomization.yaml. By default, this generates a hash-suffixed secret
that looks something like `cloudflare-api-token-1o3fji`. The
ClusterIssuer failed to find the secret because of the hash suffix. The
fix is to not generate the suffix. This does mean we'll have to delete
and recreate the secret if the API token changes, because otherwise the
change may not be picked up by cert-manager.