k8s: fix issue with potential 502s during redeployment.
See links in comments. Basically, when a pod shuts down during a
redeployment, Kubernetes may tell the pod to terminate before it's
removed from the load balancer, which can lead to failed requests. The
workaround is to add a delay to the pod termination process to give it
time for changes to the load balancer to propagate.