k8s: fix 404s for static assets being cached.
Bug: when a new deployment is being rolled out, if a user lands on a new
pod, they may send a request for one of the new JS/CSS files to an old
pod. This resulted in a 404 response, which would be incorrectly cached
by Cloudflare. This in turn broke Javascript, since Cloudflare would
return the cached 404 response for the JS file forever.

Fix: don't cache 404 responses for static assets.